Ask sim variants filename instead of directory
Now instead of asking for the directory with the base folder within it, the simulation variants Excel file can be selected. This way specific simulation series can be saved in several Excel files, which can be selected specifically for calculation.
Also the simulation series folder now also has the filename of the selected sim variants Excel file in it.