Enable passing an array for start values
Until now a single starting value could be passed to the optimization (the starting value was then turned into an array with the same value all over). Now an array with a prediction of the heating power can be passed, which ressembles more the probable output of the optimization, thus needing less time to compute.