Add workaround for Anutz* parameter
The parameters Anutz1 to Anutz4 within the dck files are set by the formulas max([56,214],1) to max([56,217],1). For an unknown reason, this formula causes the parameters to be 0, resulting in a division by 0 error.
As a workaround, the parameters are set with the value REFAREA of the respective zone, which is read from the b17/b18 file.