Improve simulation success check
Now the simulation success check works for any number of simulated hours. The simulation start and stop can be changed in the simulation variant excel file. Those values are now  used to determine the expected length of the output file (simulation was successfull, if the outputfile has the expected length).